refactor: move all code from main.go to startup.go
